Desktop/p.i.sistemafinanceiro/backend/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
psycopg2-binary
python-dotenv
flask-migrate
pytest
reportlab
pandas
